[FIX] P13nDialog: Primary action is now emphasized in warning dialog
The Problem: In the conditions panel when the selected condition
is between (BT) and we add only one of the values and press OK
there is a warning message that values are missing.

In this warning dialog the primary action button (Fix) is not emphasized.

The Solution: Now we emphasize the primary action.

BCP: 2380074541
Change-Id: I4205b419f1a12b7da4f7d7629a8c31bdd9ace694