[FIX] sap.m.DatePicker: adjust value help icon settings
Issue:
- The value help icon was reachable via screen reader in
VPC mode even when the component is disabled. In this case
the announcements are confusing to the users as there is
lacking information about the disabled state.

Solution:
- The value help icons are docorativce as per specification
for all desktop devices. On mobile devices and tables those
icons are reachable via assistive technologies.
This solution includes all date and time controls plus the
sap.m.StepInput control.

BCP: 2370075198
JIRA: FIORITECHP1-23241
Change-Id: I690f0d17a891a8db9bcf624662cfa60a86db3dc9