[INTERNAL] Demo Kit: Better handling of EoCP
Problem:
By default we expect the EoCP values to have the following format "quarter/year".
However, recently we have seen that some EoCP values are not in this format. (e.g. "To Be Determined")
This leads to some versions not being displayed in the version dropdown.

Solution:
Implement a better handling of the EoCP in the Demo Kit.

Jira: BGSOFUIPIRIN-6179
Change-Id: I3da2fc8f87e1c17b8ebc7656e8a2e77850ff3b42