[FIX] OPA: Exclude static area controls from Component parent check
Problem: When Dialog is opened and it is placed in a Component, which is
placed in a busy Control (child of another Component), OPA cannot
interact with the Dialog, as it goes through the Interactive matcher.
This matcher now (after recent change) checks all parents, including the
parent of the inner Component --> in this case, it is a busy Control.
That is why the Dialog appear to be not interactable.
Note: While this is a legit change --> controls should be non
interactable, when placed in a busy parent up in the tree, Dialogs are
placed in the static area and are actually interactable, so we need to
exclude them in this case.

Solution: Exclude controls, placed in the static area, when checking the
parent tree above the Component.
Note: Previously check has been stopped at the Component. It did not go
upward the parent tree.

BCP: 2380100969
Change-Id: I72e6fc38402f85d3671480975b10d8e1f2439e5e