[INTERNAL][FIX] Theming: Accept empty string as a theme name
Previously the UI5 Core accepted "" as a theme name and
defaulted to "base".
This change reenables this behavior and prevents the UI from
being completely unstyled.

Previous to this change a "Maximum call stack size exceeded"
error was produced during the startup of the Core when an
empty string was given as the theme name.

Additionally a bug when writing the theme-roots has been fixed,
where we accidentally reused an empty object as default value
for a configuration access.
This lead to old and new theme-roots being identical and no
"change" event being fired.

Change-Id: I8ac5c99ce2a2baf42d09d0c21abf915ed01e1873