[INTERNAL]sap.ui.rta: Outline Service - createdControls
The outline service is extended to include created controls IDs on the
extension point info.

Also, the view ID was not being properly retrieved for XML Processing in
the change handler. This was also fixed on this change.

Finally, controls can be created before the registration of the extension
point - this change also takes this case into consideration.

JIRA: TEAM42-287
Change-Id: I2ef438213f9fde99d831fda519d70fc8529f9f59