[FEATURE] Field: support data types on additionalValue
If the additionalValue has a different data type than String the output should be formatted using this data type. Also the Input should be parsed using this data type. If a ValueHelp is used, items selected from the ValueHelp should be also shown in the right format and put in the internal format into the bound model.
For FilterFields a property is needed to set the additional data type as there is no additionalValue property to bind too.

Change-Id: I5b790a01f26ce89363c21f1c41e76aef78002260
JIRA: CPOUIFDENVER-1928