[FIX] theming/ThemeManager: Avoid multiple CSS loading
IncludeLibraryTheme should only initially add a library
CSS to the DOM in order to add a library CSS to the
theming lifecycle therefore we skip adding the CSS in
case there is already a link tag with the corresponding
ID.
All other changes to the CSS links should be done by
updateThemeUrl.
Also updateThemeUrl should ensure that we always compare
the fully resolved href with the link href.

BCP: 002075129500005027552023
Change-Id: I7bb3ef1729521d5d26f294e1976ffa780aab47f8