[INTERNAL] m.Table: Focus redirect shouldn't happen for draggable rows
 - Although there is no issue found in Firefox it seems like Chrome is
not starting the drag operation if the focus changes on the mouse down.
 - With the new keyboard handling the cells are focusable via keyboard
but they shouldn't be focusable via mouse. Therefore mouse interaction
to the cell with tabindex=-1 was forwarding the focus to the row and
this focus redirect does not allow dragstart event to be firend in
Chrome.
 - In this change the tabindex gets removed on mouse down to get allow
dragging not from the cell but row.

BCP: 2380108138

Change-Id: I5807c3c45f23f881101dc7512bfdbc7ec6adfdb6