[INTERNAL] Theming: Support for default theme and fallback themes
Adds support for a default theme and calculates a valid fallback theme.

Two main scenarios are tackled:

1. No "theme" configuration is set
Previously, the default theme was set to "base".
In case no theme configuration is given, this results in a broken UI
with just the bare minimum of visual support.
With this change we establish a predetermined default theme, set to the
latest available supported theme.

2. An invalid standard theme is set
Invalid standard themes are:
  a) themes that have been deprecated and removed
  b) themes that are not yet supported in the current version

For this "rel-1.114" branch change, the latest default theme is "sap_horizon".

JIRA: CPOUI5FRAMEWORK-112
JIRA: CPOUI5FRAMEWORK-500
BCP: n/a
CR-Id: 002075125800003562432023
Change-Id: I301b79539221b4b0db7b9b1fd3f4bc9647ad1e9e