[FIX] sap.m.PlanningCalendar: Fix navigation skipping
Issue: When there is a max date set, navigation with < and > arrow keys
will sometimes skip a week to set the max day as the first day.

Solution: No longer force the max day to be set ad the first day in the
row. Also update nav buttons after rendering since they would appear as
disabled when they shouldn't be.

BCP: 002075129400005568382023
Change-Id: Ie952818aebf802099f74a123e7c0eaf18d895962