[INTERNAL] AdaptFiltersPanel: disable vertical scrolling when used in sap.m.p13n.Popup
The AdaptFiltersPanel provides its own scrolling capabilites.
With a recent refactoring, the scrolling of the outer dialog has been enabled which interfers with the scrolling of the panel itself, visually limitating the amount of items that the user can work with (although he scan still search and work with them).

Therefore a central logic has been implemented that the AdaptFiltersPanel when used in a p13n.Popup will always use its own scrolling by introducing an optional interface method #getVerticalScrolling in the sap.m.p13n.IContent interface.

BCP: 2380115464
Change-Id: Ia8696209bb7b1708e920c957acee07297c050a68