[INTERNAL][FIX] Table: getContextByIndex returns null
After a previous change, the method returned always undefined. This was done to fix inconsistent return values (either null or undefined) in Table and its subclasses.
However, the previously documented return value was null. This change makes sure that always null is returned if no context is found at the given index.

Change-Id: Ic3b081459a26731ba415b78a964fa242dc2abe4b
BCP: 2380115766