[FEATURE] sap.uxap.ObjectPageLayout: introduce beforeNavigation event
Problem: We use an ObjectPageLayout with icon tab bar (no anchor bar).
Some sections have a local edit area. When being in edit mode with
pending changes and the user tries to navigate to another section, we
ask the user whether to save or discard the pending changes or to cancel
the navigation. This logic is currently not working as there is no
suitable event, which can be used from application side to cancel the
user navigation.

Solution: Introduce a new beforeNavigation event, which can be used from
applications to cancel the user navigation, until the user makes a
decision whether he will continue with editting the currently selected
Section or he will stop editting and go to the newly selected Section.

JIRA: BGSOFUIPIRIN-6176
Change-Id: Ibb92d4a5f90b59ebec87d5df317938f1be08c026