[FIX] sap.m.TileContainer: Remove link to ACC testing page
Issue:
The TileContainer has several ACC-relevant issues, but
lacks a visual specification and seems to lack designer
support, as it is deprecated. We have received an internal
issue BCP issue reported by a tester that cannot be
properly addressed under the circmstances.

Solution:
To avoid testers sending duplicate tickets that cannot be
resolved without design support, we are removing the link
to the test page.

BCP: 2370011673
Change-Id: I0aefca236a1b5a48e45efec831625976f3671f2a