[FIX] sap.ui.unified.FileUploader: enable drop over the browse button
Issue:
- Drag and drop over the browse button was causes
the default browser action to get triggered.
The reason behind this behavior is that the button
element is displayed over the input type file field.
It was made this way so the hover and busy indication
effects are triggered over the browse button.

Solution:
- The dragged files are attached to the input type file
field programmatically when a drag and drop operation
is executed over the browse button.

BCP: 002075129400006011772023
CR-Id: 002075125800003978332023
Change-Id: I660d23c05bf0c90324228dbb73f52f4eaca4a2a0