[INTERNAL]sap.ui.fl: CompVariant - default view always favorite
If an end user sets a view to default, and later it is removed from the
favorite views by a key user, that view is still loaded on start for the
end user but they can't see it in their list of views - which looks like
an error and is counter-intuitive.

This change fixes that behavior for the CompVariant, always setting it
as favorite if it is the default view. For the FLVariant another change
will be prepared.

JIRA: TEAM42-294
Change-Id: I182f5aa025c3fd3cef30b237e4f1ae13c33e27b5