[INTERNAL]sap.ui.fl: FlVariant - default view always favorite
If an end user sets a view to default, and later it is removed from the
favorite views by a key user, that view is still loaded on start for the
end user but they can't see it in their list of views - which looks like
an error and is counter-intuitive.

This change fixes this behavior for the FlVariant, always setting it
as favorite if it is the default view. For the CompVariant there is
another change.

JIRA: TEAM42-294
Change-Id: Ie223079457c1538c52369d9149dd5052a94d1ad4