[INTERNAL]sap.ui.core: Modifier moveAggregation method
Until now, moving a control from an aggregation to another was done
calling "removeAggregation" and then "insertAggregation". However, since
these two methods are asynchronous, the control was without a parent
between the method calls. This could lead to problems in the case of
e.g. bindings with filters.

To enable the execution of a synchronous remove+insert combination,
both the XML and JS ControlTreeModifiers have a "moveAggregation" method
now.

JIRA: TEAM42-284
Change-Id: I12195f584062600bee1d9f32a77c17f4e5508fa7