[INTERNAL][FIX] GridTable: Retrieve element also in complex ID cases
Change 5923088 fixed an issue in the AutoRowMode in coombination with Safari. It used .querySelector to retrieve the elements by ID. This works for simple IDs (like in tests, etc.), but not in complex cases (ID's have a dot).
Therefore, instead of using .querySelector, we will use document.getElementById instead.

BCP: 2380130060
Change-Id: I06bc1db53aabaa1890421c4929cb3c4ad89ced4e