[FIX] sap.ui.test.Opa5: Tweaked the AutoWaitParams sample
Problem:
When this test is being executed using Karma Test Runnner in some
browser versions the test fails.

Solution:
We are adding "autoWait" flag in `iStartMyAppInAFrame` function which
gives more time for the application to startup and stabilizes tests
for slow-loading applications.

This parameter is false by default, regardless of the global "autoWait"
value, to prevent issues in existing tests.

BCP: 2380111983
Change-Id: Ife20fa822beeb540ec1e66a34a3cd6f8414553a6