[INTERNAL] messaging (tests): refactor tests using xx-handleValidation
Tests that rely on the legacy configuration `xx-handleValidation`
are modified to use the new APIs `Messaging.registerObject` and
`Messaging.unregisterObject` instead.

Copies of the original tests are kept in a separate test file which
is marked as using legacy APIs. It will be removed in a future version
of UI5.

Change-Id: Ibea44dd0a726b636bd4048a91683b9d4a3956f9f