[FIX] sap.ui.documentation: Restore 'additionalDownloadFiles' functionality for samples
When defining a UI5 sample which should be visible in the UI5 demokit it
is necessary to maintain a manifest.json on the root level of the sample.
This manifest.json have to include also an array under
'sap.ui5/config/sample/files' maintaining a list of files showed in the
sample code view in the demokit. Furthermore, it was possible to define
files which should not be visible in demokit but being part of the
downloaded artefact by specifying these files in another array under
'sap.ui5/config/sample/additionalDownloadFiles'.

Change-Id: I0163b73d1e6f96ea90df1fa8816d0ed91fb6f385