[INTERNAL] Table: Deprecate "plugins" aggregation
Following the standard established by  sap.m.plugins, plugins of sap.ui.table should now also be added to the "dependents" aggregation.
This also respects the fact that non-selection plugins don't need a dedicated "plugins" aggregation.
This is the first step in the unification of the plugin concepts in an
attempt to simplify usage for consumers.

To get a certain plugin, sap.m.plugins.PluginBase.getPlugin can be used.

PS17: Resolve merge conflicts and add tests

Change-Id: I5602b96952b3657d1ff576e46a7811151b7c147c