[INTERNAL] Stabilize and move QUnit tests
Ensure, that test expectation in XMLView.qunit.js is
checked at a time where the precoditions are fullfilled.
Move QUnit test from CoreLock.qunit.js to UIArea.qunit.js
Test was not really testing the intended behavior. Adjusted the test
and moved to UIArea where the lock functionality belongs to.

BCP: N/A
CR-Id: 002075125900002380752023
Change-Id: I1b367d4c7bf9e00c9dfe9d873f66fda3004842b4