[FEATURE] QUnit: Allow coverage watermarks configuration via frontend
With this change, default `watermarks` can
be overridden by providing an HTML data attribute
to the `qunit-coverage-istanbul.js` script or extending the
`coverage` property in `testsuite.qunit.js` files.
Watermarks' structure in `testsuite.qunit.js` is the same as in the official
documentation of `@ui5/middleware-code-coverage`: https://github.com/SAP/ui5-tooling-extensions/tree/main/packages/middleware-code-coverage#usage

This feature relies on `@ui5/middleware-code-coverage` (https://github.com/SAP/ui5-tooling-extensions/pull/145) available with version `1.1.0`. If you use an earlier `@ui5/middleware-code-coverage` version, the `qunit-coverage-istanbul.js` produces a coverage report and an error will be logged in the console.

Changes:
- `_setupAndStart.js`: Reads `coverage`
property from testsuite and writes data
attributes to the script tag on the (HTML) page
- `qunit-coverage-istanbul.js`: Reads data
attributes on the page, checks for a
configuration and sends a config request
to the middleware.

JIRA: CPOUI5FOUNDATION-709
Change-Id: Iec687b94266378596e6240fb70f41c5cb2de15c8