[INTERNAL][FIX]sap.ui.fl: CompVariant dirty after save
If a change is done on a CompVariant inside Key User Adaptation and then
a save is triggered on Key User Adaptation without saving the variant
first, the change in the variant is saved but the SVM control stays
with the dirty flag.

This fix removes the dirty flag from the control when save takes place.

BCP: 2380103905
Change-Id: Ide92e938fde3dad837ebe4797d00124865a0bf0b