[FIX] sap.f.DynamicPage: Back tab nav fixed
When tabbing back through focusable content, sometimes the focus hits
elements under the fixed header, when it is in snapped mode.
Setting scroll-padding-top on the container, with a value equal to
header height, we make sure that the top scroll position hits on the
visible part of the content.

BCP: 2370052810
Change-Id: I7ceb31ea1d92c4d87867d46f9214b411529b242f