[INTERNAL] JSDoc: allow enum definitions using `const`
The JSDoc toolkit classifies 'enum' not as a 'kind' of its own. Instead,
symbols of a different kind (e.g. 'member' or 'constant') are
additionally flagged with 'symbol.isEnum = true'.

The UI5 plugin and template so far only expected symbols of kind
'member' to be flagged with 'isEnum'. With the support of ES6+ syntax,
using a 'const' declaration also makes sense when defining enums.
JSDoc creates a symbol of kind 'constant' in that case, which needs
to be understood by the UI5 plugin and template.

The recently introduced core/message/MessageType now uses a 'const'
declaration, as originally intended.

Change-Id: I4fbc347148aaa52373963ccc2ec04f242d66ee62