[INTERNAL]sap.ui.fl: Comment on Flex Support Tool breakpoint line
Added a comment to mark which line is being used by the Flex Support
Tool for adding breakpoints to prevent that the tool stops working due
to code adjustments.

If the line needs to be changed, the tool should also be adjusted: it
should be extended to consider the new code (but still look for the
previous code as well so it works with older versions).

Change-Id: I7aac630d5314964bb320d898a92783608d46ce96