[FIX] MDCTable: Sort invisible property with ODataV4
If analytics is enabled in an MDC Table of type "Table" with an odata.v4.TableDelegate, sorting by a property that is not part of a visible column causes a backend error.
To avoid this error, the sorting of such properties is ignored (the sorter is not added to the binding info).

Change-Id: Ied7d1ff96e5076c627f2a7a696ed79ea32b81cdd
BCP: 2380131026