[INTERNAL][FIX] sap.ui.fl: Don't modify history on iframe src change
Avoid overriding the iframe src after it was initially set to not
push further history updates from e.g. parameter changes.
With this change the window.location is replaced directly in the
embedded window.
This does not affect user-initiated navigation inside the embedded site,
which will still add entries to the outer history as intended.

CR-Id: 002075125900002715502023
BCP: 002075129400006817922023
Change-Id: I383532affb592ada8fbabdb1235eeb3f4f859953