[INTERNAL] sap.f.DynamicPageTitle: Focus of hidden items prevented
Problem:
Items from the hidden area of the sap.f.DynamicPageTitle
enter the tab-chain if they have "visibility: visible"
explicitly set in their own CSS

Solution:
Since there is always the possibility for external CSS
to specify the problematic "visibility: visible" CSS rule,
fixed by ensuring this rule is overriden. Also tried
the expectedly more performance-optimal solution to set
"dislay: none" to the entire aria that we want to
exclude from the tab-chain, but discarded (in favour of
"visibility: visible") because it conflicts with RTA
requirements.

BCP: 2370087153
Change-Id: I70ed5c6bc77db0a287adce44ee60ac8118752275