[INTERNAL][FIX] Table: Initialization if bound on init
If a subclass binds the rows on init, the table wants to get the computed row counts from the row mode instance that does not exist yet.

The default row mode instance is now created if it is accessed before settings are applied. This can lead to the situation that the initial values are not correct (different than the ones provided in the settings), and therefore, for example, the initial request is sent with suboptimal values.

Actually the subclass should solve this issue, but because row modes are still private, a temporary solution needs to be provided in the table.

Change-Id: If25d773ed1df07a6f0b06fadd8e754c10a6666e6
BCP: 2370086821