[INTERNAL] sap.uxap, generic testsuite: Exclude sap.uxap.BlockBase from MemoryLeak.qunit
The BlockBase asynchronously creates a View.
The asynchronous creation combined with destruction of the BlockBase
leads to a timing and orchestration issue in the generic MemoryLeak test.

The UI5 framework does not (yet) provide a method of
orchestration for asynchronous destroy, which could be captured in
the generic tests.

The control needs to be exclude from the MemoryLeak.qunit to stabilize
test execution.

BCP: 2380127584
Change-Id: I5fcfe6fa66c0749e2ca73c3bb642450a454db47e