[INTERNAL] sap.m.table.Util: refer to core.Icon via an own dependency
The `getSelectAllPopover` method referred to the `core.Icon` control
via the core library. But such an approach relies on globals (which
won't work in future versions of UI5) and only works when the Icon
class has been loaded before or by loading it on demand using sync
loading (lazy stub).

A future-proof approach is to refer to `core.Icon` as a module of its
own and loading it dynamically.

Change-Id: I46875a1625ca9e0b20bd9dad6c00964bc8f6ea76