[INTERNAL] sap.tnt.SideNavigation: Suppress hover effect after selecting a child item
On a collapsed SideNavigation, after a child item is selected from the popover,
the hover effect on the item will be suppressed for the next interaction.
This results in the item being focused as usual, but the effect is
stopped to avoid distracting and potentially confusing the user.
The effect is restored again on the next interaction

Additionally adjusted the scrollbar to remove its square corner when the control is
expanded and there is a scroll bar showing.

JIRA: BGSOFUIRODOPI-3102
Change-Id: I52fb9c60a5d0e1439f30c66f45b25afcae028936