[INTERNAL] DemoKit: Disable some files for editing
Problem: Some files does not make sense to be edited. However, they are
used for running the sample, when downloaded.

Solution: Still displaying all files, included in the manifest.json, as
it is useful for the user to be able to review them. However, make only
.js and .xml files editable. The rest of the files are displayed in
read-only mode (MessageStrip is also shown to indicate that the files is
read-only).
Additionally, provide a way for each sample to make additional files
editable, through configuration in the manifest.json:
additionalEditableFiles array can be used. All files, included in it,
will be editable in the samples.

JIRA: BGSOFUIPIRIN-6222
Change-Id: I52190dda331cda44de01fd29c9c45802e99ca9bd