[INTERNAL] core.ScrollBar (test): stabilize "Scroll Huge Scrollbars"
This is a follow-up to I60017469d80368e42c8293c8a9c1bb774fe17293
(commit 9c7c193e6707d8212b1c192fb6cd6356bd8ecb89).

Only a sync rendering guaranteed that onAfterRendering was executed
before the onscroll event. With async rendering, onscroll might execute
first and modified the just set scrollPosition to an unexpected value.

As a short-term fix, sync rendering is restored. But as async rendering
is used in productive scenarios, the implementation of the ScrollBar
needs to be revised.

BCP: 2370089789
Change-Id: I5944b9c07b98d67bf8b71b135bfc759040efc662