[INTERNAL] Field: initially create only a few Tokens
Enable ManagedObjectModel to bind the "last x items".
Initiall only bind the last Tokens, only if focus goes to field bind all tokens. This speeds up initial rendering.
On focussing Token the binding is only updated if there are really more tokens than alredy bound. In this case the binding is updated with a delay to support the press on the delete-icon. (Here the click event is fired async.)

Adapt OPA tests to not use the Press-event on the delete-icons as this would lead to a focus event on the non-focusable icon.

Change-Id: I2e305b01c5b1d98ff244b081c8cc1a81044b2e2a
JIRA: CPOUIFDENVER-2004