[FIX] sap.uxap.ObjectPageLayout: Prevent error on mobile
Problem: When on mobile and a Section is removed from the
ObjectPageLayout and then destroyed with parameter "KeepDom" (used from
binding for keeping already existing DOMRefs and reusing them), an error
in console is thrown.

Solution: The problem happens when the second last Section is removed
(so there are no any other Sections after it to be reused from binding).
It happens that in _updateScreenHeightSectionBasesAndSpacer method we
are still working with the already removed and destroyed Section (called
from onBeforeRendering), when its DOMRef is still there. As there is a
DomRef, whole logic is executed and on a mobile device we are searching
for a parent, which in this case does not exists --> that's why an error
is thrown. (If the Section is not the second last, there will be a
parent and an error is not thrown).

BCP: 2380126677
Fixes: https://github.com/SAP/openui5/issues/3828
Change-Id: I60cac0bb909a8e2053f71dbeb8ac8a96b93c46a4