[INTERNAL] sap.ui.fl: remove .Component from reference in FlexObject constructor
Old changes were created with an additional .Component in the reference.
Since this is no longer the case, this additional part is now removed.
This way the reference in the change metadata is always the proper reference

- removed the reference as parameter from the VariantsMap DataSelector,
as the parent already has that parameter. Now the reference is taken from the first change

- removed the .Component in all the js files (except for some tests that still have functionality about this)

Change-Id: I4ac51ee41d5109e95bdc9856bbdb4cfbb5b3edb5