[INTERNAL] sap.ui.unified.FileUploader: Introduces a new FileUploaderHelper
* Change prompted by the need to remove the use of global names in version 2.x
By creating a new helper file and porting the helper logic to the FileUploader we have the ability to replace global names with hard dependencies in version 2.x without causing a circular dependency.
This change is intended to remain part of version 1.x where it is still possible for a FileUploader to be instantiated from the commons library and we need to know where to load internal controls from.
For version 2.x, the FileUploader can only be created from the unified library and hard dependencies to the m library can be used.
- This change serves only as a preparation for the easier removal of global names. The removal itself occurs in this change - 5982734.

This change and this - 5982734 is strongly related.
the second change(5982734) will be submitted with version 2.x and will not be part of version 1.x
The creation of this change(5982734) is required to be added to legacy-free transform and wait for the release of version 2.x

JIRA: BGSOFUIBALKAN-7005
Change-Id: I4454669d6bcf010cccf16276cb569b9fce049f50