[INTERNAL] sap.m.SegmentedButton: Legacy-free UI5 adjustments
Remove uses of deprecated 'select' event in controls and tests.

*Note* that the 'buttons' aggregation and 'selectedButton' association
are deprecated and supposed to be removed, but they are still being used
internally.

As agreed, the legacy-free UI5 transformer will remove the
defaultAggregation and change the 'buttons' aggregation from deprecated
to hidden (in a separate change). The 'selectedButton' association will
also be changed to hidden.

JIRA: BGSOFUIBALKAN-7168
Change-Id: I6960c1e4d282186cff696cfe7fc6c6d019fcb35e