[FIX] ResourceBundle: Use Accept-Language: * for empty locale ("")
The Accept-Language header must not have an empty value.
When the client does not have any locale preferences it should send "*"
as value in order to let the server decide about the locale.
Note that this is different from not setting the header at all, which
will let the browser set it based on browser / OS settings.

Adds missing test cases and adjust mocking of existing case via
deprecated jquery.sap.resources module.
Removes unnecessary beforeEach in ResourceBundle QUnit module and moves
stub to the only test where it is required.

BCP: 2370100866
Change-Id: Id1c7c557f2ebb6bebdee25883bfbd08071c4f9a4