[INTERNAL] sap.ui.layout.Splitter: Don't freeze areas sizes after user resize
Problem:
Whenever some of the bars is dragged, it results in "px" sizes to its adjacent areas.
After the splitter is resized, through browser resize for example, the splitter remains with blank space.

Solution:
Whenever some of the bars is dragged, set the size of the preceding area in the same unit as its initial unit.
If the area is sized with "auto", then set the size of the subsequent area.
For example "x%" will be resized to "y%", "px" will be resize to "px" and "rem" will be "rem".
"auto" areas are not converted to another unit, unless there are 2 adjacent "auto" areas. In this case they will be converted to "%".

BCP: 2370086984
Change-Id: Ib336a110dcfa82fa3743d7be7479114b8b103b75