[INTERNAL] Form: put FormHelper as module in layout library
To remove the ObjectPath logic.
FormHelper is now an own module what holds the library specific implementations. Depending ob what library is loaded it provided the library-specific implementation.

As the old FormHelper is used in SmartForm environment too, the old FormHelper needs to be removed after the SmartForm is adoped.

JIRA: CPOUIFDENVER-1905

Change-Id: I785b2660e9a9ee8375226850582dca8418f2e742