[FIX] sap.m.OverflowToolbar: Fix overflowing controls with 100% width
Problem: When there is only one Control, placed in the OverflowToolbar,
and it has 100% width (coming from CSS class, not set inline), the
Control goes in the overflow menu, while it should not (it just needs to
take the whole content space of the OverflowToolbar).

Solution: Replace left margin of the first child of the OverflowToolbar
with left-pading on the OverflowToolbar itself. Now we have similar
left/right margins.

BCP: 2370093601
Change-Id: I1eb4109cd85ac0c1a1999ef34be3ea7d00e4c5c3