[INTERNAL] m.GrowingEnablement: Set the focus after binding update
 - When the focus is on the growing list control, the rendering after
the binding update may remove the focus, so the focus should be set
again after the rendering is complete. If the list control is busy at
that moment, the busy handler disables all events, including focus
events, and this does not allow ItemNavigation to process the focus
event and put it on the previously focused item. In this change, after
the busy state is set to false, if the focus is still on the root on the
list control instead of items then we trigger the focus event to let the
ItemNavigation processes it.

BCP: 2380107627

Change-Id: Ic7c729ad05e02750c28fd66592a2a2a967c1099d