[INTERNAL] sap.ui.rta (tests): fix controller implementations
In UI5 2.0, controllers are expected to be modules of their own, they're
no longer retrieved via a global name. Tests that implement a controller
as fixture need to expose that controller like a module, either by using
a dedicated module or by using an embedded, named sap.ui.define call.

Change-Id: Idb62c146fae5fedceb2f12c43987983bd57e1801