[INTERNAL] CopyProvider: Handling of unsecure context
The application should check whether it is executed in secure context before adding the CopyProvider
plugin, otherwise the plugin will be activated and throw an error once the user tries to copy.
In the MDCTable the plugin and the copy button are not added in case of unsecure context.

JIRA: CPOUIFTEAMB-2226
Change-Id: Id895347f19f88fce5634a199dabdfae6b4a8eb01