[FIX] DemoKit: Version button visibility fixed
Previously, the "Version" switch button was conditional based on
whether the framework version was a snapshot.
Now, it is consistently visible regardless of the framework version.

Additionally, we've enhanced the clarity of the additional
object status text. It will now be displayed only when the framework
encounters difficulties while requesting information about other versions.

Change-Id: I2706f53c3d3bd14e0b32a2309bfac723476c6410