[INTERNAL][FIX] sap.m.Toolbar: Fix failing visual tests
Forms - overrides of margin-left for the first child are not
needed anymore (as they come now by default from Toolbar). Override
margin-left of first child, when the old value of the Toolbar was used
(0.25rem). Adjust left-padding where needed. Remove the
:not(.sapMTBActive) selector, as it has more weight and it prevails over
forms' selectors.
OPL - removing the :not(.sapMTBActive) selector.
Active Toolbar - fixed with a new selector, specific for the active
Toolbars, which need left padding (with 0.25rem less than the others), as they have a hidden button as a first element, so the first visual element is not actually the first child (having left margin 0), but it has left margin (0.25rem - standard between each item).
Additionaly - if the first child (or even more first children) are invisible, the left spacing would not be OK. That's why we are now using sapMBarChildFirstChild class, to apply the 0 left-margin to the first VISIBLE child of OFT/TB.

BCP: N/A
Change-Id: I8d81e19095748087df4d9675dcc85465726fa597