[INTERNAL] sap.f.FlexibleColumnLayout: Enhance reveal/conceal mechanism
Problem: Due to the current mechanism of showing/hiding FLC's columns,
in some cases, scroll position of the ObjectPageLayout is changed during
transitioning from one layout to another, which leads to a wrong
selected Section/SubSection.

Solution: Enhance the reveal/conceal mechanism of FLC in a way that
scroll and resize events are no longer fired for columns that are
getting hidden/shown. Their content is now pinned (with fix width)
during the transition and also the ResizeHandler is suspended, so that
when they get finally shown (if previously hidden) no changes in width
are detected --> no resize callbacks are executed.

JIRA: BGSOFUIPIRIN-6219
Change-Id: Ib7ffcd5f4c3873749c08588bbf6d5f634c111007