[INTERNAL] sap.m.ProgressIndicator: Horizon fixes
When animating from X percent to 100%, the remaining part of the
ProgressIndicator is not visible at all.
Now it is visible, until getting overlapped by the value part.
In previous themes, the ProgressIndicator has different background (in
horizon its transparent) and/or border, that's why this issue is not
present there. However, in horizon it is visible.

JIRA: BGSOFUIPIRIN-6258
Change-Id: I591859521660b16397f91893ee30d6dfa1c7c7cc