[FEATURE] Deprecation of ManagedObject.create
The public API ManagedObject.create is deprecated.

Consequently the usage of object literals as aggregation content is also
deprecated, in case the object type is either:
* given via string through the "Type" property (capital 'T')
* derived from the aggregation type (string of class name in dot notation)

We additionally introduce the a new runtime metadata property: 'defaultClass'
as an alternative for use cases in which object literals are given as
binding templates or directly as aggregation content.

The 'defaultClass' allows control developers to specify a default
class for an aggregation. If specified, the default class will be
instantiated instead of the derived aggregation type.

Rationale:
In case of string usage, the given class types must be resolved via
the global namespace, which will not be supported in 2.0 anymore.
Additionally this can lead to the retrieval of a lazy-stub from the global namespace
and thus also causing synchronous requests to JS resources.

JIRA: CPOUI5FRAMEWORK-536
Change-Id: Ic7160aefac660008410f59c8facd5018dab28580