[FIX] sap.m.Input: Reopen Suggestion Popover when new data is bound
When having a single item bound to the tabular suggestions and when
typing new data is received with a single item again, mutation observer
is not called  and only Item's properties are updated. Input does not get
notified for that change, therefore we need to listen on aggregation update
as well.

BCP: 2370107738
Change-Id: If04c932e940de58986c1b8d4472c656f6f920ebe