[INTERNAL] Deprecate sap.ui.layout.GridHelper
The GridHelper was exposed via a globals and specifically set for
sap.m and sap.ui.commons.
With the sap.ui.commons library being deprecated and sap.m (as well
as sap.ui.layout) don't expose a special class handling.
The library specific code for handling a style class in the Grid
is thus not needed in v2.0 anymore.

JIRA: CPOUI5FRAMEWORK-317
Change-Id: Ib7ec1f6259f9dc0133dd30182ba6941d9aeb90ac