[FIX] sap.f.DynamicPageHeader: Aligned with visual specification
Problem: When the pin button of the DynamicPageHeader is pressed/toggled
the icon is not changing. It's supposed to change to similar filled icon.

Solution: On horizon themes we change the icon of the pin button when
it's toggled.

BCP: 2370081951
BCP: 2370103521
Change-Id: I56fba22e1fe95749ee0a498af37bd8d99f102052
(cherry picked from commit 9851fa26f844d02372bdfb23efa7ce4b061b1a44)
CR-Id: 002075125800001514432024
SNOW: DINC0114366