[INTERNAL] Table: Deprecated "columnSelect" event
This was forgotton when deprecating the old column menu implementation in 1.117 (replaced with sap.m.table.columnmenu.*).

The event is only fired before the sap.ui.unified menu is opened, at the same time the "columnMenuOpen" event of Column is fired, which has been deprecated. After the old menu implementation is removed (UI5 2), the columnSelect event is no longer fired.

BCP: n/a
CR-Id: 002075125800004558502023
Change-Id: Id1677b1b9ab1c502ff1b11d40c582a30678d3452