[INTERNAL] Component: Load model classes async. before instantiation
Previous to this change, all manifest model classes were sync. required during the
constructor call of a Component.
In best-practice scenarios, this did not pose a problem since
the model classes should be part of the dependency closure of the Component.
In this case the requireSync only executed the module factory of
an already preloaded module.

In UI5 2.0, requireSync is not an option anymore and all model classes
must be required asynchronously via sap.ui.require.

In the same best-practice scenario, not single request needs to be sent
because the model classes are already part of the dependencies of the Component.

This change also ensures, that ALL model classes are loaded async.
including those of models that have not been marked as "preload: true".
The time of instantiation however is unchanged.
Late models that are not marked as "preload: true" will still be instantiated
during a Components constructor call.

Additionally some legacy codings have been marked as deprecated.

JIRA: CPOUI5FRAMEWORK-210
Change-Id: I3d6f00f276f24c72b573a574c384e82ab5a69e0c