[INTERNAL] CellSelector: Remove cell selection on selection change
Whenever a row is selected or deselected (the selection changes), then the cell selection should be cleared.

This avoids a clash between row selection and cell selection, which lead to unexpected behaviour and heavily confused
end users and was deemed as a bug.

Additionally, after discussions with UX, this change will also solve the issue, that the cell selection
was perceived to be "active for too long", as it persisted it through additional row selections and other actions.
This change will fix up this behaviour, matching users expectations.

To improve UX and avoid future this should be merged.

BCP: 2370123901
CR-Id: 002075125800004886842023
Change-Id: Ia1da3a9868d3cf28087137ab93b060772196db5a
(cherry picked from commit 9ad0cdb90fa9d7c4b62482cb618e95e9d1d231f2)