[INTERNAL] CopyProvider: HTML MIME-Type support with ClipboardItem
 - text/plain: contains the simple format, each property in a column
extracted.
 - text/html: contains the visual representation, combined data
properties.

In Excel text/html is preferred. If the user copies from a UI5 Table and
pastes to Excel then the user will see what is shown in the UI5 Table.
The user can also use "Paste Special" and select the "Text" option to
see the simplified version which can be used for editing and then
pasting back to the original UI5 Table.

In Firefox, if the ClipboarItem support is not enabled with the
"dom.events.asyncClipboard.clipboardItem" option then we only use the
text/plain(simplified) version.
In Mobile, even though the ClipboarItem is supported, there is no Mobile
SpreadSheet application that supports "Paste Special" therefore we again
only use the text/plain(simplified) version.

Change-Id: I77fd6b7a2866eeec522ec29bf62a362f17f4d7ef
CR-Id: 002075125900003579162023
BCP: 2380137834
BCP: 2370145549
BCP: 2380134371