[INTERNAL] CopyProvider: HTML MIME-Type support with ClipboardItem
 - text/plain: contains the simple format, each property in a column
extracted.
 - text/html: contains the visual representation, combined data
properties.

In Excel text/html is preferred. If the user copies from a UI5 Table and
pastes to Excel then the user will see what is shown in the UI5 Table.
The user can also use "Paste Special" and select the "Text" option to
see the simplified version which can be used for editing and then
pasting back to the original UI5 Table.

In Firefox, if the ClipboarItem support is not enabled with the
"dom.events.asyncClipboard.clipboardItem" option then we only use the
text/plain(simplified) version.
In Mobile, even though the ClipboarItem is supported, there is no Mobile
SpreadSheet application that supports "Paste Special" therefore we again
only use the text/plain(simplified) version.

Jira: CPOUIFTEAMB-2249
Jira: CPOUIFTEAMB-2260
BCP: 2370145549

Change-Id: I3f05f286af31e3d0d5f38df867174ee2796b1368